`lookupConformance` request is not cached and constraint solver performs a lot of them for the same type
(i.e. during disjunction solving), let's try to cache previously performed requests to see whether additional
memory use is worth the performance benefit.


<!--
Before merging this pull request, you must run the Swift continuous integration tests.
For information about triggering CI builds via @swift-ci, see:
https://github.com/apple/swift/blob/main/docs/ContinuousIntegration.md#swift-ci

Thank you for your contribution to Swift!
-->