@xedin 2023-02-24
@swift-ci please test compiler performance