@xedin 2023-02-25
@swift-ci please test compiler performance