@xedin 2023-02-26
@meg-gupta @eeckstein I think one of the recent changes to CSE optimization pass introduced a non-deterministic crash, this time “new” built but “old” crashed (that is the version of the compiler without any modifications introduced by this PR), last time “new” crashed this way. Could you please take a look?