@xedin 2023-02-27
@swift-ci please test compiler performance