@xedin 2023-02-27
# Summary for main full

No regressions above thresholds

- [Debug-batch](#debug-batch-6aca360d7b497479)
    - [debug-batch brief](#debug-batch-brief-6aca360d7b497479)
    - [debug-batch detailed](#debug-batch-detailed-6aca360d7b497479)
- [Release](#release-6aca360d7b497479)
    - [release brief](#release-brief-6aca360d7b497479)
    - [release detailed](#release-detailed-6aca360d7b497479)



<a name="debug-batch-6aca360d7b497479">

# Debug-batch

<a name="debug-batch-brief-6aca360d7b497479">

## debug-batch brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 91,550,459,123,743 | 89,972,141,531,967 | -1,578,317,591,775 | -1.72% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (1)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
LLVM.NumLLVMBytesOutput | 2,107,523,492 | 2,107,566,772 | 43,280 | 0.0%
</details>

<a name="debug-batch-detailed-6aca360d7b497479">

## debug-batch detailed
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 91,550,459,123,743 | 89,972,141,531,967 | -1,578,317,591,775 | -1.72% :white_check_mark:
Sema.LookupConformanceInModuleRequest | 76,719,941 | 50,860,339 | -25,859,602 | -33.71% :white_check_mark:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (309)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 3,195,914 | 3,195,959 | 45 | 0.0%
AST.ImportSetCacheMiss | 703,811 | 703,818 | 6 | 0.0%
AST.ImportSetFoldHit | 353,813 | 353,819 | 6 | 0.0%
AST.ImportSetFoldMiss | 349,998 | 349,999 | 0 | 0.0%
AST.ModuleShadowCacheHit | 8,451 | 8,451 | 0 | 0.0%
AST.ModuleShadowCacheMiss | 2,515 | 2,515 | 0 | 0.0%
AST.ModuleVisibilityCacheHit | 18,011,655 | 18,011,654 | 0 | -0.0%
AST.ModuleVisibilityCacheMiss | 150,307 | 150,306 | 0 | -0.0%
AST.NumASTBytesAllocated | 51,331,421,767 | 51,330,874,522 | -547,245 | -0.0%
AST.NumASTScopeExpansions | 7,233,456 | 7,233,447 | -9 | -0.0%
AST.NumASTScopeLookups | 8,044,751 | 8,044,593 | -158 | -0.0%
AST.NumDecls | 303,831 | 303,831 | 0 | 0.0%
AST.NumDependencies | 408,350 | 408,350 | 0 | 0.0%
AST.NumIncrementalDependencies | 21,346 | 21,346 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 293,404 | 293,404 | 0 | 0.0%
AST.NumLocalTypeDecls | 2,438 | 2,438 | 0 | 0.0%
AST.NumModuleLookupClassMember | 6,820 | 6,820 | 0 | 0.0%
AST.NumModuleLookupValue | 88,712,186 | 88,713,002 | 816 | 0.0%
AST.NumObjCMethods | 73,374 | 73,374 | 0 | 0.0%
AST.NumOperators | 1,022 | 1,022 | 0 | 0.0%
AST.NumPrecedenceGroups | 172 | 172 | 0 | 0.0%
AST.NumReferencedDynamicNames | 159 | 159 | 0 | 0.0%
AST.NumReferencedMemberNames | 12,040,327 | 12,040,339 | 12 | 0.0%
AST.NumReferencedTopLevelNames | 1,158,719 | 1,158,721 | 2 | 0.0%
AST.NumSourceBuffers | 337,466 | 337,466 | 0 | 0.0%
AST.NumSourceLines | 10,454,984 | 10,454,980 | -4 | -0.0%
AST.NumSourceLinesPerSecond | 4,946,567 | 4,961,027 | 14,460 | 0.29%
AST.NumTotalClangImportedEntities | 1,689,028 | 1,689,525 | 496 | 0.03%
Frontend.MaxMallocUsage | 848,499,387,312 | 847,578,727,784 | -920,659,528 | -0.11%
Frontend.NumProcessFailures | 31 | 31 | 0 | 0.0%
IRGen.IRGenRequest | 26,420 | 26,420 | 0 | 0.0%
IRGen.OptimizedIRRequest | 0 | 0 | 0 | 0.0%
IRGen.SymbolObjectCodeRequest | 0 | 0 | 0 | 0.0%
IRModule.NumGOTEntries | 258,051 | 258,051 | 0 | 0.0%
IRModule.NumIRAliases | 208,756 | 208,756 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 7,345,703 | 7,345,709 | 6 | 0.0%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 3,883,672 | 3,883,676 | 4 | 0.0%
IRModule.NumIRGlobals | 3,778,777 | 3,778,783 | 6 | 0.0%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 89,940,024 | 89,940,106 | 82 | 0.0%
IRModule.NumIRNamedMetaData | 131,016 | 131,016 | 0 | 0.0%
IRModule.NumIRValueSymbols | 7,719,118 | 7,719,128 | 10 | 0.0%
LLVM.NumLLVMBytesOutput | 2,107,523,492 | 2,107,566,772 | 43,280 | 0.0%
Parse.IDEInspectionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 336,168 | 336,168 | 0 | 0.0%
Parse.NumIterableDeclContextParsed | 681,650 | 681,646 | -4 | -0.0%
Parse.ParseAbstractFunctionBodyRequest | 573,787 | 573,787 | 0 | 0.0%
Parse.ParseMembersRequest | 524,875 | 524,871 | -4 | -0.0%
Parse.ParseSourceFileRequest | 330,254 | 330,254 | 0 | 0.0%
Parse.ParseTopLevelDeclsRequest | 330,254 | 330,254 | 0 | 0.0%
SILGen.ASTLoweringRequest | 27,685 | 27,685 | 0 | 0.0%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 2,921,094 | 2,921,098 | 4 | 0.0%
SILModule.NumSILGenGlobalVariables | 120,991 | 120,991 | 0 | 0.0%
SILModule.NumSILGenVtables | 36,437 | 36,437 | 0 | 0.0%
SILModule.NumSILGenWitnessTables | 156,320 | 156,320 | 0 | 0.0%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 3,437,811 | 3,437,815 | 4 | 0.0%
SILModule.NumSILOptGlobalVariables | 122,803 | 122,803 | 0 | 0.0%
SILModule.NumSILOptVtables | 36,738 | 36,738 | 0 | 0.0%
SILModule.NumSILOptWitnessTables | 191,707 | 191,707 | 0 | 0.0%
SILOptimizer.ExecuteSILPipelineRequest | 108,238 | 108,238 | 0 | 0.0%
SILOptimizer.LoweredSILRequest | 0 | 0 | 0 | 0.0%
Sema.ABIMembersRequest | 266,790 | 266,788 | -1 | -0.0%
Sema.AbstractGenericSignatureRequest | 129,553 | 129,582 | 29 | 0.02%
Sema.AccessLevelRequest | 15,511,082 | 15,509,374 | -1,707 | -0.01%
Sema.ActorIsolationRequest | 6,019,807 | 6,019,919 | 111 | 0.0%
Sema.AllMembersRequest | 99,990 | 99,990 | 0 | 0.0%
Sema.AnyObjectLookupRequest | 234 | 234 | 0 | 0.0%
Sema.ApplyAccessNoteRequest | 3,766,093 | 3,765,926 | -167 | -0.0%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 41,422 | 41,424 | 2 | 0.0%
Sema.AttachedPropertyWrapperTypeRequest | 24,825 | 24,829 | 4 | 0.02%
Sema.AttachedPropertyWrappersRequest | 19,158,867 | 19,157,722 | -1,144 | -0.01%
Sema.AttachedResultBuilderRequest | 1,642,345 | 1,642,349 | 4 | 0.0%
Sema.BodyInitKindRequest | 75,303 | 75,303 | 0 | 0.0%
Sema.BreakTargetRequest | 5,633 | 5,633 | 0 | 0.0%
Sema.CXXNamespaceMemberLookup | 0 | 0 | 0 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 325,726 | 325,726 | 0 | 0.0%
Sema.CheckDistributedFunctionRequest | 80 | 80 | 0 | 0.0%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 10,066 | 10,066 | 0 | 0.0%
Sema.CheckInconsistentSPIOnlyImportsRequest | 53,356 | 53,356 | 0 | 0.0%
Sema.CheckInconsistentWeakLinkedImportsRequest | 10,066 | 10,066 | 0 | 0.0%
Sema.CheckRedeclarationRequest | 1,993,667 | 1,993,667 | 0 | 0.0%
Sema.ClangCategoryLookupRequest | 0 | 0 | 0 | 0.0%
Sema.ClangDirectLookupRequest | 13,483 | 13,483 | 0 | 0.0%
Sema.ClangRecordMemberLookup | 13,483 | 13,483 | 0 | 0.0%
Sema.ClassAncestryFlagsRequest | 136,652 | 136,657 | 4 | 0.0%
Sema.ClosureEffectsRequest | 121,343 | 121,343 | 0 | 0.0%
Sema.ClosureHasExplicitResultRequest | 43,880 | 43,880 | 0 | 0.0%
Sema.CollectOverriddenDeclsRequest | 11,564,257 | 11,564,581 | 323 | 0.0%
Sema.CompareDeclSpecializationRequest | 681,968 | 682,178 | 210 | 0.03%
Sema.CompilerPluginLoadRequest | 0 | 0 | 0 | 0.0%
Sema.ConditionalRequirementsRequest | 742,987 | 737,880 | -5,107 | -0.69%
Sema.ConformanceHasEffectRequest | 51 | 51 | 0 | 0.0%
Sema.ConstantValueInfoRequest | 0 | 0 | 0 | 0.0%
Sema.ContinueTargetRequest | 1,975 | 1,975 | 0 | 0.0%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 22,529 | 22,529 | 0 | 0.0%
Sema.CustomAttrTypeRequest | 3,396 | 3,396 | 0 | 0.0%
Sema.CustomRefCountingOperation | 0 | 0 | 0 | 0.0%
Sema.CxxRecordSemantics | 0 | 0 | 0 | 0.0%
Sema.DefaultAndMaxAccessLevelRequest | 105,770 | 105,770 | 0 | 0.0%
Sema.DefaultArgumentExprRequest | 91,786 | 91,786 | 0 | 0.0%
Sema.DefaultArgumentInitContextRequest | 844 | 844 | 0 | 0.0%
Sema.DefaultArgumentTypeRequest | 20,107 | 20,107 | 0 | 0.0%
Sema.DefaultDefinitionTypeRequest | 7,637 | 7,637 | 0 | 0.0%
Sema.DefaultTypeRequest | 413,101 | 413,101 | 0 | 0.0%
Sema.DerivativeAttrOriginalDeclRequest | 0 | 0 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 60,856,105 | 60,854,733 | -1,372 | -0.0%
Sema.DirectOperatorLookupRequest | 1,972,018 | 1,972,115 | 97 | 0.0%
Sema.DirectPrecedenceGroupLookupRequest | 763,517 | 763,517 | 0 | 0.0%
Sema.DistributedModuleIsAvailableRequest | 67 | 67 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 6,417,781 | 6,417,894 | 113 | 0.0%
Sema.EnumRawTypeRequest | 35,771 | 35,771 | 0 | 0.0%
Sema.EnumRawValuesRequest | 20,360 | 20,360 | 0 | 0.0%
Sema.ExistentialConformsToSelfRequest | 26,621 | 26,616 | -5 | -0.02%
Sema.ExistentialRequiresAnyRequest | 15,370 | 15,370 | 0 | 0.0%
Sema.ExpandAccessorMacros | 1,123,946 | 1,123,950 | 4 | 0.0%
Sema.ExpandConformanceMacros | 102,613 | 102,613 | 0 | 0.0%
Sema.ExpandMacroExpansionDeclRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMemberAttributeMacros | 35,398,367 | 35,397,055 | -1,312 | -0.0%
Sema.ExpandPeerMacroRequest | 3,600,535 | 3,600,543 | 8 | 0.0%
Sema.ExpandSynthesizedMemberMacroRequest | 908,432 | 908,427 | -5 | -0.0%
Sema.ExtendedNominalRequest | 605,379 | 605,379 | 0 | 0.0%
Sema.ExtendedTypeRequest | 185,823 | 185,822 | -1 | -0.0%
Sema.ExternalMacroDefinitionRequest | 0 | 0 | 0 | 0.0%
Sema.FragileFunctionKindRequest | 2,354,693 | 2,354,697 | 4 | 0.0%
Sema.FunctionOperatorRequest | 17,583 | 17,583 | 0 | 0.0%
Sema.GenericParamListRequest | 6,102,098 | 6,103,278 | 1,180 | 0.02%
Sema.GenericSignatureRequest | 3,160,527 | 3,160,376 | -151 | -0.0%
Sema.GetDestructorRequest | 42,987 | 42,987 | 0 | 0.0%
Sema.GetDistributedActorArgumentDecodingMethodRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorIDPropertyRequest | 72 | 72 | 0 | 0.0%
Sema.GetDistributedActorImplicitCodableRequest | 117 | 117 | 0 | 0.0%
Sema.GetDistributedActorInvocationDecoderRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedActorSystemPropertyRequest | 56 | 56 | 0 | 0.0%
Sema.GetDistributedActorSystemRemoteCallFunctionRequest | 46 | 46 | 0 | 0.0%
Sema.GetDistributedRemoteCallArgumentInitFunctionRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedRemoteCallTargetInitFunctionRequest | 19 | 19 | 0 | 0.0%
Sema.GetDistributedTargetInvocationDecoderDecodeNextArgumentFunctionRequest | 22 | 22 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordArgumentFunctionRequest | 20 | 20 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordErrorTypeFunctionRequest | 11 | 11 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordReturnTypeFunctionRequest | 16 | 16 | 0 | 0.0%
Sema.GetDistributedTargetInvocationResultHandlerOnReturnFunctionRequest | 2 | 2 | 0 | 0.0%
Sema.GetDistributedThunkRequest | 80 | 80 | 0 | 0.0%
Sema.GetImplicitSendableRequest | 307,090 | 307,665 | 575 | 0.19%
Sema.GetRuntimeDiscoverableAttributes | 2,098,768 | 2,098,768 | 0 | 0.0%
Sema.GetSourceFileAsyncNode | 142 | 142 | 0 | 0.0%
Sema.GlobalActorAttributeRequest | 7,670,607 | 7,670,680 | 73 | 0.0%
Sema.GlobalActorInstanceRequest | 1,116 | 1,116 | 0 | 0.0%
Sema.HasCircularInheritedProtocolsRequest | 17,042 | 17,042 | 0 | 0.0%
Sema.HasCircularRawValueRequest | 20,716 | 20,716 | 0 | 0.0%
Sema.HasDefaultInitRequest | 91,110 | 91,112 | 2 | 0.0%
Sema.HasDynamicCallableAttributeRequest | 27,581 | 27,581 | 0 | 0.0%
Sema.HasDynamicMemberLookupAttributeRequest | 521,627 | 521,625 | -2 | -0.0%
Sema.HasImportsMatchingFlagRequest | 655,187 | 655,187 | 0 | 0.0%
Sema.HasIsolatedSelfRequest | 13,310,355 | 13,310,650 | 294 | 0.0%
Sema.HasMemberwiseInitRequest | 41,070 | 41,072 | 2 | 0.0%
Sema.HasMissingDesignatedInitializersRequest | 30,652 | 30,652 | 0 | 0.0%
Sema.HasUserDefinedDesignatedInitRequest | 91,120 | 91,122 | 2 | 0.0%
Sema.IDEInspectionFileRequest | 0 | 0 | 0 | 0.0%
Sema.InferredGenericSignatureRequest | 219,359 | 219,356 | -3 | -0.0%
Sema.InheritedDeclsReferencedRequest | 6,295,248 | 6,294,607 | -641 | -0.01%
Sema.InheritedProtocolsRequest | 766,335 | 766,317 | -17 | -0.0%
Sema.InheritedTypeRequest | 427,028 | 427,087 | 58 | 0.01%
Sema.InheritsSuperclassInitializersRequest | 39,252 | 39,252 | 0 | 0.0%
Sema.InitKindRequest | 166,301 | 166,303 | 2 | 0.0%
Sema.InterfaceTypeRequest | 19,962,524 | 19,961,979 | -545 | -0.0%
Sema.IsABICompatibleOverrideRequest | 280,611 | 280,611 | 0 | 0.0%
Sema.IsAccessorTransparentRequest | 511,389 | 511,389 | 0 | 0.0%
Sema.IsActorRequest | 2,044,449 | 2,044,920 | 471 | 0.02%
Sema.IsCallAsFunctionNominalRequest | 6,205 | 6,205 | 0 | 0.0%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDeclRefinementOfRequest | 54,982 | 54,976 | -6 | -0.01%
Sema.IsDefaultActorRequest | 58,198 | 58,198 | 0 | 0.0%
Sema.IsDistributedActorRequest | 2,286,713 | 2,287,210 | 497 | 0.02%
Sema.IsDynamicRequest | 2,825,746 | 2,825,752 | 6 | 0.0%
Sema.IsFinalRequest | 4,236,490 | 4,237,180 | 689 | 0.02%
Sema.IsGetterMutatingRequest | 700,416 | 700,416 | 0 | 0.0%
Sema.IsImplicitlyUnwrappedOptionalRequest | 3,493,526 | 3,493,578 | 52 | 0.0%
Sema.IsMoveOnlyRequest | 649,517 | 649,523 | 5 | 0.0%
Sema.IsObjCRequest | 2,620,793 | 2,620,626 | -167 | -0.01%
Sema.IsSafeUseOfCxxDecl | 0 | 0 | 0 | 0.0%
Sema.IsSetterMutatingRequest | 621,156 | 621,156 | 0 | 0.0%
Sema.IsSingleValueStmtRequest | 323,924 | 323,924 | 0 | 0.0%
Sema.IsStaticRequest | 1,459,608 | 1,459,607 | -1 | -0.0%
Sema.LazyStoragePropertyRequest | 2,423 | 2,423 | 0 | 0.0%
Sema.LocalDiscriminatorsRequest | 463,787 | 463,787 | 0 | 0.0%
Sema.LookupAllConformancesInContextRequest | 2,443,816 | 2,444,101 | 285 | 0.01%
Sema.LookupInModuleRequest | 8,276,243 | 8,278,161 | 1,917 | 0.02%
Sema.LookupInfixOperatorRequest | 69,635 | 69,637 | 2 | 0.0%
Sema.LookupPostfixOperatorRequest | 74 | 74 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 29,583 | 29,583 | 0 | 0.0%
Sema.LookupPrefixOperatorRequest | 352 | 352 | 0 | 0.0%
Sema.MacroDefinitionRequest | 0 | 0 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 1,108 | 1,108 | 0 | 0.0%
Sema.ModuleImplicitImportsRequest | 10,066 | 10,066 | 0 | 0.0%
Sema.ModuleLibraryLevelRequest | 351 | 351 | 0 | 0.0%
Sema.ModuleQualifiedLookupRequest | 3,204,692 | 3,206,768 | 2,075 | 0.06%
Sema.NamedLazyMemberLoadSuccessCount | 33,823,844 | 33,822,737 | -1,106 | -0.0%
Sema.NamingPatternRequest | 241,094 | 241,097 | 3 | 0.0%
Sema.NeedsNewVTableEntryRequest | 948,694 | 948,694 | 0 | 0.0%
Sema.NumAccessorBodiesSynthesized | 417,421 | 417,421 | 0 | 0.0%
Sema.NumAccessorsSynthesized | 512,121 | 512,119 | -2 | -0.0%
Sema.NumConformancePathsRecorded | 845,696 | 845,637 | -59 | -0.01%
Sema.NumConformancesDeserialized | 5,106,877 | 5,106,158 | -719 | -0.01%
Sema.NumConstraintScopes | 52,792,737 | 52,792,800 | 62 | 0.0%
Sema.NumConstraintsConsideredForEdgeContraction | 1,860,790 | 1,860,790 | 0 | 0.0%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 0 | 0 | 0 | 0.0%
Sema.NumDeclsDeserialized | 84,602,571 | 84,593,918 | -8,653 | -0.01%
Sema.NumDeclsTypechecked | 2,826,208 | 2,826,208 | 0 | 0.0%
Sema.NumLazyIterableDeclContexts | 12,092,765 | 12,091,429 | -1,336 | -0.01%
Sema.NumLazyRequirementSignatures | 1,019,068 | 1,018,912 | -155 | -0.02%
Sema.NumLazyRequirementSignaturesLoaded | 694,913 | 694,808 | -105 | -0.02%
Sema.NumLeafScopes | 39,751,331 | 39,751,294 | -37 | -0.0%
Sema.NumRequirementMachineCompletionSteps | 23,822,829 | 23,821,998 | -831 | -0.0%
Sema.NumRequirementMachineUnifiedConcreteTerms | 1,035,573 | 1,036,261 | 688 | 0.07%
Sema.NumRequirementMachines | 1,987,994 | 1,988,465 | 471 | 0.02%
Sema.NumTypesDeserialized | 24,097,729 | 24,101,916 | 4,187 | 0.02%
Sema.NumUnloadedLazyIterableDeclContexts | 8,582,114 | 8,581,019 | -1,094 | -0.01%
Sema.ObjCInterfaceAndImplementationRequest | 752,203 | 752,203 | 0 | 0.0%
Sema.OpaqueReadOwnershipRequest | 447,818 | 447,818 | 0 | 0.0%
Sema.OpaqueResultTypeRequest | 862 | 862 | 0 | 0.0%
Sema.OperatorPrecedenceGroupRequest | 1,021 | 1,021 | 0 | 0.0%
Sema.OverriddenDeclsRequest | 5,344,099 | 5,345,193 | 1,094 | 0.02%
Sema.ParamSpecifierRequest | 1,955,098 | 1,955,098 | 0 | -0.0%
Sema.PatternBindingEntryRequest | 692,807 | 692,810 | 3 | 0.0%
Sema.PatternTypeRequest | 812,298 | 812,301 | 3 | 0.0%
Sema.PolymorphicEffectKindRequest | 977,005 | 977,009 | 4 | 0.0%
Sema.PolymorphicEffectRequirementsRequest | 60 | 60 | 0 | 0.0%
Sema.PreCheckResultBuilderRequest | 1,041 | 1,041 | 0 | 0.0%
Sema.PreCheckReturnStmtRequest | 505,807 | 505,807 | 0 | 0.0%
Sema.PrimaryAssociatedTypesRequest | 7,231 | 7,231 | 0 | 0.0%
Sema.PrimarySourceFilesRequest | 10,066 | 10,066 | 0 | 0.0%
Sema.PropertyWrapperAuxiliaryVariablesRequest | 2,050,284 | 2,050,284 | 0 | 0.0%
Sema.PropertyWrapperBackingPropertyTypeRequest | 24,825 | 24,829 | 4 | 0.02%
Sema.PropertyWrapperInitializerInfoRequest | 596,684 | 596,686 | 2 | 0.0%
Sema.PropertyWrapperLValuenessRequest | 2,912 | 2,912 | 0 | 0.0%
Sema.PropertyWrapperMutabilityRequest | 696,097 | 696,097 | 0 | 0.0%
Sema.PropertyWrapperTypeInfoRequest | 626 | 626 | 0 | 0.0%
Sema.ProtocolDependenciesRequest | 593,384 | 593,387 | 3 | 0.0%
Sema.ProtocolRequirementsRequest | 73,382 | 73,380 | -1 | -0.0%
Sema.ProtocolRequiresClassRequest | 43,973 | 44,051 | 78 | 0.18%
Sema.ProvideDefaultImplForRequest | 11,564,257 | 11,564,581 | 323 | 0.0%
Sema.QualifiedLookupRequest | 8,383,415 | 8,383,513 | 98 | 0.0%
Sema.RangeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.RenamedDeclRequest | 4,225 | 4,225 | 0 | 0.0%
Sema.RequirementRequest | 175,260 | 175,260 | 0 | -0.0%
Sema.RequirementSignatureRequest | 749,737 | 749,695 | -41 | -0.01%
Sema.RequiresOpaqueAccessorsRequest | 1,789,894 | 1,789,896 | 2 | 0.0%
Sema.RequiresOpaqueModifyCoroutineRequest | 449,554 | 449,554 | 0 | 0.0%
Sema.ResolveEffectiveMemberwiseInitRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveImplicitMemberRequest | 790,770 | 790,702 | -67 | -0.01%
Sema.ResolveMacroRequest | 29,891 | 29,891 | 0 | 0.0%
Sema.ResolveProtocolNameRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeEraserTypeRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeRequest | 5,032,036 | 5,032,029 | -6 | -0.0%
Sema.ResultBuilderTypeRequest | 466,125 | 466,129 | 4 | 0.0%
Sema.ResultTypeRequest | 1,103,608 | 1,103,607 | -1 | -0.0%
Sema.RootAndResultTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.RootTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.SPIGroupsRequest | 7,334,374 | 7,334,801 | 427 | 0.01%
Sema.ScopedImportLookupRequest | 810 | 810 | 0 | 0.0%
Sema.SelfAccessKindRequest | 910,737 | 910,743 | 5 | 0.0%
Sema.SelfBoundsFromGenericSignatureRequest | 67,205 | 67,205 | 0 | 0.0%
Sema.SelfBoundsFromWhereClauseRequest | 1,757,558 | 1,758,094 | 535 | 0.03%
Sema.SemanticAvailableRangeAttrRequest | 5,391 | 5,391 | 0 | 0.0%
Sema.SemanticUnavailableAttrRequest | 148,904 | 148,904 | 0 | 0.0%
Sema.SetterAccessLevelRequest | 245,443 | 245,443 | 0 | 0.0%
Sema.SimpleDidSetRequest | 2,869,492 | 2,868,646 | -846 | -0.03%
Sema.SpecializeAttrTargetDeclRequest | 3,084 | 3,084 | 0 | 0.0%
Sema.StorageImplInfoRequest | 2,049,746 | 2,049,750 | 4 | 0.0%
Sema.StoredPropertiesAndMissingMembersRequest | 62,670 | 62,670 | 0 | 0.0%
Sema.StoredPropertiesRequest | 381,669 | 381,671 | 2 | 0.0%
Sema.StructuralRequirementsRequest | 54,983 | 55,046 | 63 | 0.12%
Sema.StructuralTypeRequest | 1,934 | 1,934 | 0 | 0.0%
Sema.SuperclassDeclRequest | 516,027 | 516,026 | -1 | -0.0%
Sema.SuperclassTypeRequest | 67,113 | 67,113 | 0 | 0.0%
…